Check if the domain object is set (mounted is done) before trying to destroy the mutable

+++ src/plugins/flexibleLayout/components/frame.vue
@@ -115,7 +115,7 @@ export default {
         this.dragGhost = document.getElementById('js-fl-drag-ghost');
     },
     beforeDestroy() {
-        if (this.domainObject.isMutable) {
+        if (this.domainObject !== undefined && this.domainObject.isMutable) {
             this.openmct.objects.destroyMutable(this.domainObject);
         }
